Added flags to LD cgo to make CUDA lib be used and corrected type return for device count

+++ go/cuda.go
@@ -3,6 +3,6 @@ package torch
 // #include "gotorch.h"
 import "C"
 
-func CudaDeviceCount() uintptr {
+func CudaDeviceCount() C.ulong {
 	return C.cuda_getDeviceCount()
 }
